chore: 🤖 sendo new palette file to @mamba/styles
