--- packages/configs/postcss/includes/colors.js
@@ -1,60 +1,0 @@
-const colors = {
-  seashell: '#f1f1f1',
-  neutral900: '#181D25',
-  neutral800: '#252C37',
-  neutral700: '#353E4B',
-  neutral600: '#5C6A7F',
-  neutral500: '#7D8BA1',
-  neutral400: '#A3ADBD',
-  neutral300: '#CBD2DC',
-  neutral200: '#DDE2E9',
-  neutral100: '#EAECF1',
-  neutral50: '#F3F5F7',
-  green900: '#004033',
-  green800: '#005C41',
-  green700: '#00734D',
-  green600: '#008E5A',
-  green500: '#00A868',
-  green400: '#0FCC7D',
-  green300: '#42EC9A',
-  green200: '#A3F5CC',
-  green100: '#DEF7EA',
-  green50: '#E8FEF3',
-  blue700: '#00699E',
-  blue600: '#007FBF',
-  blue500: '#0091D9',
-  blue400: '#16A9F3',
-  blue300: '#79CAF2',
-  blue200: '#B3E6FF',
-  blue100: '#D6F2FF',
-  purple700: '#583294',
-  purple600: '#6A43AA',
-  purple500: '#7D54C0',
-  purple400: '#9D77D9',
-  purple300: '#AA8DD9',
-  purple200: '#D8C7F5',
-  purple100: '#F0E6FF',
-  red700: '#CC0700',
-  red600: '#E41811',
-  red500: '#F5342E',
-  red400: '#FF5752',
-  red300: '#FF8E8A',
-  red200: '#FFD3D1',
-  red100: '#FFECEB',
-  orange700: '#C14D05',
-  orange600: '#E75C06',
-  orange500: '#F9701B',
-  orange400: '#FF914C',
-  orange300: '#FBA874',
-  orange200: '#FEDDC8',
-  orange100: '#FFEEE0',
-  yellow700: '#CC7A00',
-  yellow600: '#F79502',
-  yellow500: '#FAA60A',
-  yellow400: '#FDBB2C',
-  yellow300: '#FFDB80',
-  yellow200: '#FFECB3',
-  yellow100: '#FFF4D1',
-};
-
-module.exports = colors;
